+++ src/main/java/com/yohanii/lostandfound/domain/chatting/RoomRepository.java
@@ -52,4 +52,16 @@ public class RoomRepository {
                 || (room.getMember().getId().equals(partnerId) && room.getPartnerId().equals(memberId)))
                 .findFirst();
     }
+
+    public void deleteAll(List<Long> roomIds) {
+        findAll().stream()
+                .filter(room -> roomIds.contains(room.getId()))
+                .forEach(em::remove);
+    }
+
+    public long getRoomCount() {
+        return ((Number) em.createQuery("select count(r) from Room r")
+                .getSingleResult())
+                .longValue();
+    }
 }