+++ src/main/java/com/yohanii/lostandfound/domain/member/MemberAuth.java
@@ -1,0 +1,6 @@
+package com.yohanii.lostandfound.domain.member;
+
+public enum MemberAuth {
+
+    ADMIN, MEMBER
+}