+++ src/main/java/com/yohanii/lostandfound/web/interceptor/AdminInterceptor.java
@@ -1,0 +1,30 @@
+package com.yohanii.lostandfound.web.interceptor;
+
+import com.yohanii.lostandfound.domain.member.Member;
+import com.yohanii.lostandfound.domain.member.MemberAuth;
+import com.yohanii.lostandfound.web.SessionConst;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.servlet.HandlerInterceptor;
+
+@Slf4j
+public class AdminInterceptor implements HandlerInterceptor {
+
+    @Override
+    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
+
+        HttpSession session = request.getSession();
+        if (session == null || session.getAttribute(SessionConst.LOGIN_MEMBER) == null || session.getAttribute(SessionConst.LOGIN_MEMBER) instanceof Member member) {
+            Member loginMember = (Member) session.getAttribute(SessionConst.LOGIN_MEMBER);
+            if (loginMember.getAuth() != MemberAuth.ADMIN) {
+                log.info("AdminInterceptor : MemberAuth is not ADMIN");
+                response.sendRedirect("/");
+                return false;
+            }
+        }
+
+        return true;
+    }
+}